3.4: Castor fiber
Introduce numeric to covscript

+++ include/covscript/core/definition.hpp
@@ -80,7 +80,7 @@ namespace cs {
 	template<typename _Tp> using set_t = std::unordered_set<_Tp>;
 #endif
 	using var = cs_impl::any;
-	using number = long double;
+	using number = numeric;
 	using boolean = bool;
 	using string = std::string;
 	using list = std::list<var>;
